requirements.txt
# Core dependencies for Modal deployment
modal>=0.63.0

# For local testing and visualization
requests>=2.31.0
pillow>=10.0.0
numpy>=1.26.0
matplotlib>=3.8.0

# Optional: for local SAM2 testing (not needed for Modal deployment)
# torch>=2.4.1
# torchvision>=0.19.1
# huggingface_hub>=0.25.2
# git+https://github.com/facebookresearch/sam2.git@c2ec8e14a185632b0a5d8b161928ceb50197eddc
